backend/requirements.txt
uwsgi
pylint
pyserial
